requirements.txt
cryptography
dockerspawner
jupyterhub
pyjwt
tornado
traitlets
